requirements.txt
attrs==24.3.0
certifi==2024.12.14
charset-normalizer==3.4.1
h11==0.14.0
idna==3.10
iniconfig==2.0.0
jinja2==3.1.5
markupsafe==3.0.2
outcome==1.3.0.post0
packaging==24.2
pluggy==1.5.0
pysocks==1.7.1
pytest==8.3.4
pytest-html==4.1.1
pytest-metadata==3.1.1
pyyaml==6.0.2
requests==2.32.3
selenium==4.27.1
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.28.0
trio-websocket==0.11.1
typing-extensions==4.12.2
urllib3==2.3.0
websocket-client==1.8.0
wsproto==1.2.0
